refactor: Separate extended feats from PluginMetaManager to PluginMetaProvider
