+++ src/main/java/net/kunmc/lab/teamkunpluginmanager/utils/PluginUtil.java
@@ -8,6 +8,7 @@ import org.bukkit.plugin.PluginDescriptionFile;
 import org.bukkit.plugin.PluginLoadOrder;
 import org.bukkit.plugin.java.JavaPlugin;
 import org.bukkit.plugin.java.PluginClassLoader;
+import org.jetbrains.annotations.NotNull;
 
 import java.io.File;
 import java.io.FileNotFoundException;
@@ -155,5 +156,10 @@ public class PluginUtil  // TODO: Rewrite this class
         zip.close();
 
         return desc;
+    }
+
+    public static String normalizePluginName(@NotNull String name)
+    {
+        return name.toLowerCase().replace(" ", "_");
     }
 }
